add styles for menu

+++ FoodDeliveryApi/DTO/MenuDishDTO.cs
@@ -8,7 +8,7 @@ namespace FoodDeliveryApi.Models
     {
         public int ID { get; set; }
         public string Title { get; set; }
-        public int CookingTime { get; set; }
+        public double CookingTime { get; set; }
         public decimal Price { get; set; }
     }
 }
